backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-generativeai==0.5.3
ollama==0.1.8
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.4.5
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
pydantic==2.7.1
